Fix netboot build failure with 64-bit
This fix resolves the following build failure:

| ../netboot/linux-asm-io.h: Assembler messages:
| ../netboot/linux-asm-io.h:108: Error: operand size mismatch for `in'
| Makefile:714: recipe for target 'libgrubefi_a-pci.o' failed

+++ netboot/linux-asm-io.h
@@ -105,7 +105,7 @@ extern inline void outs##s(unsigned short port, const void * addr, unsigned long
 
 __IN(b,"", char)
 __IN(w,"",short)
-__IN(l,"", long)
+__IN(l,"", int)
 
 __OUT(b,"b",char)
 __OUT(w,"w",short)
